LPS-117060 Rename useParseURL to URLParser
